+++ muna/beta/metadata/litert.py
@@ -6,7 +6,7 @@
 from pydantic import Field
 from typing import Literal
 
-from ._torch import PyTorchInferenceMetadataBase, TorchExporter
+from ._torch import PyTorchInferenceMetadataBase
 
 class LiteRTInferenceMetadata(PyTorchInferenceMetadataBase):
     """
@@ -14,9 +14,8 @@ class LiteRTInferenceMetadata(PyTorchInferenceMetadataBase):
 
     Members:
         model (torch.nn.Module): PyTorch module to apply metadata to.
-        model_args (tuple[Tensor,...]): Positional inputs to the model.
+        model_args (tuple): Positional inputs to the model.
         input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
-        output_keys (list): Model output dictionary keys. Use this if the model returns a dictionary.
     """
     kind: Literal["meta.inference.litert"] = Field(default="meta.inference.litert", init=False)
-    exporter: TorchExporter | None = Field(default=None, init=False)
+    exporter: None = Field(default=None, init=False, exclude=True)