+++ muna/beta/metadata/mlx.py
@@ -1,0 +1,35 @@
+# 
+#   Muna
+#   Copyright © 2026 NatML Inc. All Rights Reserved.
+#
+
+from pydantic import Field
+from typing import Literal
+
+from ._torch import PyTorchInferenceMetadataBase
+from .onnxruntime import OnnxRuntimeInferenceSessionMetadata
+
+class MLXInferenceMetadata(PyTorchInferenceMetadataBase):
+    """
+    Metadata to compile a PyTorch model for inference with MLX on Apple Silicon.
+
+    Members:
+        model (torch.nn.Module): PyTorch module to apply metadata to.
+        exporter (TorchExporter): PyTorch exporter to use.
+        model_args (tuple): Positional inputs to the model.
+        input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
+        optimum_config (optimum.ExporterConfig): Optimum exporter configuration. Required when `exporter` is `optimum`.
+    """
+    kind: Literal["meta.inference.mlx"] = Field(default="meta.inference.mlx", init=False)
+
+class MLXInferenceSessionMetadata(OnnxRuntimeInferenceSessionMetadata):
+    """
+    Metadata to compile an OnnxRuntime `InferenceSession` for inference with MLX on Apple Silicon.
+
+    Members:
+        session (onnxruntime.InferenceSession): OnnxRuntime inference session to apply metadata to.
+        model_path (str | Path): ONNX model path. The file must exist in the compiler sandbox.
+        external_data_path (str | Path): ONNX model external data path. This file must exist in the compiler sandbox.
+    """
+    kind: Literal["meta.inference.mlx_onnx"] = Field(default="meta.inference.mlx_onnx", init=False)
+    providers: None = Field(default=None, init=False, exclude=True)